Podcast Chapter Proposal
Podcast Chapters are a feature supported by many [podcast clients](https://stratechery.com/outline/what-podcast-player-do-you-recommend/) already however the current standard is to [embed the chapter content into the MP3 file itself](https://id3.org/id3v2-chapters-1.0). This is podcast specific content and should be moved to the podcast RSS feed directly so it can be more easily accessed and updated by podcasters.

There is some precedent for moving this into the RSS Feed. The [Simple Chapters Specification](https://podlove.org/simple-chapters/) was introduced in 2012 and I've drawn on that spec heavily for this proposal for introduction in the Podcast Namespace.