Updated podcast:value specification for custom records
This commit makes the usage of customKey and customValue more specific.

The change to the spec was made based on the known users for these
attributes (LNPay and HIVE gateway), but it is not necessarily backwards
compatible.

Why is this change neccessary?

Client applications that use the information on the podcast:value
tag must know what to with its content.

Imagine one receiver sets customValue to raw binary data encoded in
hex, and expects it to be sent in the payment as raw binary
data.

Imagine another receiver sets customValue to raw binary data encoded in
base64, and expects it to be sent in the payment as raw binary
data.

Imagine now that other receiver sets customerValue to a string, and
expects it to be sent in the payment as a string encoded in UTF-8 (which
currently known users do).

There is no way a client could know what do to with customValue (how to
decode it an re-encode it), without specific knowledge about how each
receiver uses this field.

Although this is possible, it is counter productive for client
development, as clients should have to know as little as possible
about different application specific protocols used by the receivers.