Correct file extension used for CC0
f3eeda4 (Adding license, 2022-10-14) added a copy of the plain text
version of CC0 [1]. That commit mistakenly used the .md file extension
for the plain text file that it created.

GitHub automatically assumes that .md files contain Markdown [2]. As a
result, the plain text version of CC0 would get rendered as Markdown
which made it significantly harder to read. The lists inside sections 1
and 4 would be rendered as amorphous paragraphs.

This commit changes the file extension to .txt to prevent GitHub from
making CC0 harder to read.

[1]: <https://creativecommons.org/publicdomain/zero/1.0/legalcode.txt>
[2]: <https://github.com/github/markup/blob/b2230a29592f84e459a5278275fa4da8b9a57289/README.md#markups>